refactor(gc): split the runtime slot-store helpers into barrier_store.rs (2000-line cap)
